+++ magda-web-client/src/helpers/record.js
@@ -55,6 +55,17 @@ export type DatasetPublisher = {
     publisher: Publisher
 };
 
+type CompatiblePreviews = {
+    map?: boolean,
+    chart?: boolean,
+    table?: boolean,
+    json?: boolean,
+    html?: boolean,
+    text?: boolean,
+    rss?: boolean,
+    google?: boolean
+};
+
 //aspect=dcat-distribution-strings
 export type RawDistribution = {
     id: string,
@@ -68,7 +79,8 @@ export type RawDistribution = {
             fields: Object,
             format: string,
             timeseries: boolean,
-            wellFormed: boolean
+            wellFormed: boolean,
+            compatiblePreviews: CompatiblePreviews
         }
     }
 };
@@ -99,12 +111,13 @@ export type ParsedDistribution = {
     format: string,
     downloadURL: ?string,
     accessURL: ?string,
-    updatedDate: string,
+    updatedDate: ?string,
     license: string,
     linkActive: boolean,
     linkStatusAvailable: boolean,
     isTimeSeries: boolean,
-    chartFields?: ?Object
+    chartFields?: ?Object,
+    compatiblePreviews: CompatiblePreviews
 };
 
 // all aspects become required and must have value
@@ -179,7 +192,21 @@ const defaultDistributionAspect = {
         fields: {},
         format: null,
         timeseries: false,
-        wellFormed: false
+        wellFormed: false,
+        compatiblePreviews: null
+        // Decisions to be made here on what the default should be
+        // For now the default has to be null so that it can be overriden by the
+        //  guessCompatiblePreviews output
+        // {
+        //     map: false,
+        //     chart: false,
+        //     table: false,
+        //     json: false,
+        //     html: false,
+        //     text: false,
+        //     rss: false,
+        //     google: false
+        // }
     }
 };
 
@@ -191,6 +218,61 @@ function getFormatString(aspects) {
     if (formatAspect && formatAspect["format"]) return formatAspect["format"];
     if (dcatAspect && dcatAspect["format"]) return dcatAspect["format"];
     return defaultString;
+}
+
+function guessCompatiblePreviews(format, isTimeSeries): CompatiblePreviews {
+    // Make a guess of compatible previews from the format
+    // Should be temporary before it's properly implemented
+    //  in the "visualization sleuther"
+    const compatiblePreviews = {
+        map: false,
+        chart: false,
+        table: false,
+        json: false,
+        html: false,
+        text: false,
+        rss: false,
+        google: false
+    };
+    const fmt = format.toLowerCase();
+
+    if (fmt.indexOf("csv") !== -1) {
+        compatiblePreviews.table = true;
+        compatiblePreviews.google = true;
+        compatiblePreviews.chart = !!isTimeSeries;
+    }
+    switch (fmt) {
+        case "wms":
+        case "geojson":
+        case "wfs":
+        case "csv-geo-au":
+        case "kml":
+            compatiblePreviews.map = true;
+            break;
+        case "xls":
+        case "xlsx":
+        case "doc":
+        case "docx":
+        case "pdf":
+            compatiblePreviews.google = true;
+            break;
+        case "rss":
+            compatiblePreviews.rss = true;
+            break;
+        case "json":
+            compatiblePreviews.json = true;
+            break;
+        case "text":
+        case "txt":
+            compatiblePreviews.text = true;
+            break;
+        case "htm":
+        case "html":
+            compatiblePreviews.html = true;
+            break;
+        default:
+    }
+    return compatiblePreviews;
 }
 
 export function parseDistribution(
@@ -215,6 +297,9 @@ export function parseDistribution(
     const linkStatusAvailable = Boolean(linkStatus.status); // Link status is available if status is non-empty string
     const linkActive = linkStatus.status === "active";
     const isTimeSeries = aspects["visualization-info"]["timeseries"];
+    const compatiblePreviews =
+        aspects["visualization-info"].compatiblePreviews ||
+        guessCompatiblePreviews(format, isTimeSeries);
     let chartFields = null;
 
     if (isTimeSeries) {
@@ -242,7 +327,8 @@ export function parseDistribution(
         linkStatusAvailable,
         linkActive,
         isTimeSeries,
-        chartFields
+        chartFields,
+        compatiblePreviews
     };
 }
 
@@ -305,12 +391,16 @@ export function parseDataset(dataset?: RawDataset): ParsedDataset {
                 numeric: numericFields
             };
         }
+        const format = getFormatString(distributionAspects);
+        const compatiblePreviews =
+            distributionAspects["visualization-info"].compatiblePreviews ||
+            guessCompatiblePreviews(format, isTimeSeries);
         return {
             identifier: d["id"],
             title: d["name"],
             downloadURL: info.downloadURL || null,
             accessURL: info.accessURL || null,
-            format: getFormatString(distributionAspects),
+            format,
             license:
                 !info.license || info.license === "notspecified"
                     ? "License restrictions unknown"
@@ -320,7 +410,8 @@ export function parseDataset(dataset?: RawDataset): ParsedDataset {
             linkActive: linkStatus.status === "active",
             updatedDate: info.modified ? getDateString(info.modified) : null,
             isTimeSeries: visualizationInfo["timeseries"],
-            chartFields
+            chartFields,
+            compatiblePreviews
         };
     });
     return {