+++ CHANGES.md
@@ -1,3 +1,5 @@
+## 0.0.39
+
 ## 0.0.38
 
 *   Use grid system from DTA design guide
@@ -39,6 +41,7 @@
 *   Enable homepage stories and updated homepage config
 *   Upgraded TerriaJs to 5.7.0 to fix the issue with previewing certain datasets
 *   Created `ISSUE_TEMPLATE.md` file
+*   Stopped user feedback from being duplicated
 *   Useless patch request to registry API should not trigger any event creation
 *   Allow users to select recent search item by arrow keys
 *   Add 3 more government CSW services
@@ -51,6 +54,7 @@
 *   Fixed an issue that users may see an error after page 4
 *   Fixed an issue that prevents users from searching current search text by clicking search button or pressing enter key
 *   keep distance between search box and the first story on mobile view
+*   Fixed distribution previews
 
 ## 0.0.37